Temp Fixed ui problem and improved UI
found a temporary solution for my ui problem. because of this fix i
could work on my ui without problem so i added a battlescreen,
storescreen and a partyscreen; these screen got the correct size and
template and some of the buttons but are not finished yet. i finished
startscreen. the choicescreen is almost done. I still need to make the
connection to the player script for the player stats